Kernel Module layout:
	- All modules are now part of the lib crate
	- This allows for very easy testing
	- Unforntunately I am now running into https://github.com/japaric/cortex-m-rtfm/issues/149s://github.com/japaric/cortex-m-rtfm/issues/149
	which means this project is forced back onto nightly - hopefully
	can be fixed before submission
