+++ Cargo.toml
@@ -20,7 +20,7 @@ edition = "2018"
 [dependencies]
 cortex-m = "0.5.8"
 cortex-m-rt = "0.6.7"
-cortex-m-rtfm = { version = "0.4.2" } # , features = ["timer-queue"]
+cortex-m-rtfm = { version = "0.4.2", features = ["nightly"] }
 panic-itm = "0.4.0"
 panic-semihosting = "0.5.1"
 embedded-graphics = "0.4.7"
