+++ src/application/states/mwatch.rs
@@ -1,7 +1,7 @@
 
 
 use crate::application::wm::{State, StaticState, Signal};
-use crate::Ssd1351;
+use crate::types::Ssd1351;
 use crate::system::system::System;
 
 use embedded_graphics::Drawing;
@@ -9,7 +9,7 @@ use embedded_graphics::fonts::Font6x12;
 use embedded_graphics::image::Image16BPP;
 use embedded_graphics::prelude::*;
 
-use mwatch_kernel_api::InputEvent;
+use crate::types::InputEvent;
 
 
 pub struct MWState {}
