+++ src/application/states/uop.rs
@@ -1,14 +1,14 @@
 
 
 use crate::application::wm::{State, StaticState, Signal};
-use crate::Ssd1351;
+use crate::types::Ssd1351;
 use crate::system::system::System;
 
 use embedded_graphics::Drawing;
 use embedded_graphics::image::Image16BPP;
 use embedded_graphics::prelude::*;
 
-use mwatch_kernel_api::InputEvent;
+use crate::types::InputEvent;
 
 
 pub struct UopState {}
